+++ Assignment2/Controllers/EmployeeController.cs
@@ -19,6 +19,7 @@ namespace Assignment2.Controllers
         [HttpGet]
         public async Task<ActionResult<ICollection<EmployeeDto>>> GetEmployeesByDepartmentAsync([FromRoute] string departmentId)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var employeesByDepartmentIdServiceResponse = await _employeeService.GetEmployeesByDepartmentIdAsync(departmentId);
             return employeesByDepartmentIdServiceResponse.IsSuccess ? Ok(employeesByDepartmentIdServiceResponse.Data) : NotFound(employeesByDepartmentIdServiceResponse.Message);
         }
@@ -26,6 +27,7 @@ namespace Assignment2.Controllers
         [HttpGet("{employeeId}")]
         public async Task<ActionResult<EmployeeDto>> GetEmployeeByIdAsync([FromRoute] string departmentId, [FromRoute] string employeeId)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var employeeByIdServiceResponse = await _employeeService.GetEmployeeByIdAsync(departmentId, employeeId);
             return (employeeByIdServiceResponse.IsSuccess) ? Ok(employeeByIdServiceResponse.Data) : NotFound(employeeByIdServiceResponse.Message);
         }
@@ -33,6 +35,7 @@ namespace Assignment2.Controllers
         [HttpPut("{employeeId}")]
         public async Task<ActionResult<EmployeeDto>> UpdateEmployeeByIdAsync([FromRoute] string departmentId, [FromRoute] string employeeId, [FromBody] EmployeeDto employeeDto)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var isEmployeeUpdatedServiceResponse = await _employeeService.UpdateEmployeeAsync(departmentId, employeeId, employeeDto);
             return isEmployeeUpdatedServiceResponse.IsSuccess ? Ok(employeeDto) : NotFound(isEmployeeUpdatedServiceResponse.Message);
         }
@@ -40,6 +43,7 @@ namespace Assignment2.Controllers
         [HttpPost]
         public async Task<ActionResult<EmployeeDto>> CreateNewEmployeeAsync([FromRoute] string departmentId, [FromBody] EmployeeDto employeeDto)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var isEmployeeCreatedServiceResponse = await _employeeService.CreateEmployeeAsync(departmentId, employeeDto);
             return isEmployeeCreatedServiceResponse.IsSuccess ? Ok(employeeDto) : Conflict(isEmployeeCreatedServiceResponse.Message);
         }
@@ -47,6 +51,7 @@ namespace Assignment2.Controllers
         [HttpDelete("{employeeId}")]
         public async Task<IActionResult> DeleteEmployeeByIdAsync([FromRoute] string departmentId, [FromRoute] string employeeId)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var isEmployeeDeletedServiceResponse = await _employeeService.DeleteEmployeeByIdAsync(departmentId, employeeId);
             return isEmployeeDeletedServiceResponse.IsSuccess ? Ok() : NotFound(isEmployeeDeletedServiceResponse.Message);
         }
